Refactor app structure by moving API routes to a separate module and implementing rate limiting in tests
