+++ backend/routes/__init__.py
@@ -1,0 +1,2 @@
+# All API and web routes for the Flask app are defined here.
+# Import this file in app.py and register the routes with the Flask app instance.
